% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec.R
\name{\%is\%}
\alias{\%is\%}
\title{Check if two objects are identical}
\usage{
a \%is\% b
}
\arguments{
\item{a}{an object.}

\item{b}{an object.}
}
\value{
TRUE if a and b are identical; FALSE otherwise.
}
\description{
Check if two objects are identical
}
