% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_then.R
\name{map_then_bind_cols}
\alias{map_then_bind_cols}
\title{Apply a function to a data.frame and bind the result columns to it}
\usage{
map_then_bind_cols(.df, .f)
}
\arguments{
\item{.df}{A data.frame}

\item{.f}{A function do apply on `.df`}
}
\description{
Apply a function to a data.frame and bind the result columns to it
}
