% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_then.R
\name{map_then_join}
\alias{map_then_join}
\title{Apply a function to a data.frame and join the results in it.}
\usage{
map_then_join(.df, .f, .join_by = NULL, join_function)
}
\arguments{
\item{.df}{A data.frame}

\item{.f}{A function do apply on `.df`}

\item{.join_by}{The columns to join, as passed to `left_join`}

\item{join_function}{The function used in the join}
}
\description{
Apply a function to a data.frame and join the results in it.
}
