% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_if.R
\name{mutate-if}
\alias{mutate-if}
\alias{map_if_condition}
\alias{map_if_true}
\alias{map_if_false}
\title{Apply functions depending on a condition}
\usage{
map_if_condition(.x, condition, .f_if_true = identity, .f_if_false = identity)

map_if_true(.x, condition, .f = identity)

map_if_false(.x, condition, .f = identity)
}
\arguments{
\item{.x}{An object}

\item{condition}{TRUE or FALSE}

\item{.f_if_true}{Function to apply when `condition` is `TRUE`}

\item{.f_if_false}{Function to apply when `condition` is not `TRUE`}

\item{.f}{Function to apply}
}
\value{
`.x` or a function applied to `.x`, depending on `condition`
}
\description{
Apply functions depending on a condition
}
