% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec.R
\name{vec_filter}
\alias{vec_filter}
\title{Filter a vector}
\usage{
vec_filter(.x, .f = vec_all_true)
}
\arguments{
\item{.x}{A vector.}

\item{.f}{A single predicate function or a logical vector of the same length as .x.}
}
\value{
The vector .x with (possibly) some entries deleted.
}
\description{
Filter a vector
}
