% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec.R
\name{vec_map}
\alias{vec_map}
\title{Apply a function to a vector}
\usage{
vec_map(.x, .f)
}
\arguments{
\item{.x}{A vector}

\item{.f}{A function}
}
\value{
A vector of the same size as `x`
}
\description{
Apply a function to a vector
}
